:tada: Meta Views! Render a view of your data
Allows to specify a base dataset and a possibly nested set of views on
that dataset. The labels are rendered as specified in the view and
example loading is done per view on the base dataset.

Allowing nested labels makes it neccessary that labels are always numpy arrays!
This is why a lot of tests are updated and some Dataset classes need
fixing.